+++ src/extended-dayjs.ts
@@ -1,0 +1,10 @@
+/**
+ * import dayjs from 'extended-dayjs' instead of 'dayjs' directly to ensure that extension will work correctly.
+ */
+
+import dayjs from 'dayjs'
+import utc from 'dayjs/plugin/utc'
+
+dayjs.extend(utc)
+
+export { default } from 'dayjs'
